+++ service/builder/builder.go
@@ -16,13 +16,11 @@ package builder
 
 import (
 	"flag"
-	"fmt"
 )
 
 const (
 	// flags
-	configCfg      = "config"
-	memBallastFlag = "mem-ballast-size-mib"
+	configCfg = "config"
 
 	kindLogKey        = "component_kind"
 	kindLogsReceiver  = "receiver"
@@ -34,24 +32,15 @@ const (
 )
 
 var (
-	configFile     *string
-	memBallastSize *uint
+	configFile *string
 )
 
 // Flags adds flags related to basic building of the collector application to the given flagset.
 func Flags(flags *flag.FlagSet) {
 	configFile = flags.String(configCfg, "", "Path to the config file")
-	memBallastSize = flags.Uint(memBallastFlag, 0,
-		fmt.Sprintf("Flag to specify size of memory (MiB) ballast to set. Ballast is not used when this is not specified. "+
-			"default settings: 0"))
 }
 
 // GetConfigFile gets the config file from the config file flag.
 func GetConfigFile() string {
 	return *configFile
-}
-
-// MemBallastSize returns the size of memory ballast to use in MBs
-func MemBallastSize() int {
-	return int(*memBallastSize)
 }